Updatings bonus and removing useless function

+++ srcs/bonus/main_bonus.c
@@ -18,6 +18,8 @@
 */
 
 int		g_rep;
+char	*g_read;
+int		g_sta;
 
 void	ft_prompt(void)
 {
@@ -51,12 +53,16 @@ void	sig_handler(int sig)
 	{
 		ft_putstr("\n");
 		g_rep = 130;
+		g_sta = 1;
 		ft_prompt();
 	}
-	if (sig == SIGQUIT)
+	if (sig == SIGQUIT && (!g_read || (g_read
+	&& ft_strlen(g_read) == 0)))
+		miniprintf("\b\b  \b\b");
+	else if (sig == SIGQUIT && g_read && ft_strlen(g_read) > 0)
 	{
-		kill(1, SIGINT);
 		miniprintf("Quitter (core dumped)\n");
+		kill(1, SIGINT);
 	}
 }
 
@@ -77,30 +83,42 @@ void	parse_read(char *read, t_list **env)
 	}
 	ft_lstclear(&start, *free);
 	free(read);
+	g_read = NULL;
+}
+
+void	do_read(t_list **lst_env)
+{
+	g_read = ft_checkquotes(g_read);
+	if (ft_checkread(g_read))
+		parse_read(g_read, lst_env);
+	else
+		g_rep = 512;
+	ft_prompt();
+	g_rep = (g_sta == 1 ? 130 : g_rep);
+	g_sta = 0;
 }
 
 int		main(int ac, char **av, char **env)
 {
 	t_list	*lst_env;
-	char	*read;
 
 	g_rep = 0;
 	signal(SIGINT, sig_handler);
 	signal(SIGQUIT, sig_handler);
-	if (ac > 0 && av[0])
-	{
-		lst_env = ft_ato_lst(env);
-		handle_shlvl(&lst_env);
-		ft_prompt();
-		while (get_next_line(0, &read))
-		{
-			read = ft_checkquotes(read);
-			if (ft_checkread(read))
-				parse_read(read, &lst_env);
-			ft_prompt();
-		}
-		ft_lstclear(&lst_env, *free);
-		free(read);
-		miniprintf("exit\n");
-	}
+	if (ac <= 0 || !av[0])
+		return (1);
+	lst_env = ft_ato_lst(env);
+	handle_shlvl(&lst_env);
+	ft_prompt();
+	while (1)
+		if (get_next_line(0, &g_read) == 1)
+			do_read(&lst_env);
+		else if (ft_strlen(g_read) == 0)
+			break ;
+		else
+			free(g_read);
+	ft_lstclear(&lst_env, *free);
+	free(g_read);
+	miniprintf("exit\n");
+	return (0);
 }
